latest update

+++ src/main/java/com/gtproject/librarysystem/models/Category.java
@@ -8,18 +8,20 @@ import jakarta.persistence.*;
 public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     private String name;
 
     public Category() {
     }
 
-    public Long getId() {
+    //getter setter empty const
+
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
